Create Demos
In this example, we use an infinite while (true) loop to update the time and date every second. We use DateTime.Now to get the current date and time and then print it to the console with Console.WriteLine(). After the time and date are printed, there is a delay of one second with Thread.Sleep(1000) and then the console is cleared with Console.Clear() to update the displayed information.